1. delete todolist; 2. add equals and hashCode realizations in entity-classes; 3. change numeric int id to long; 5. makee some changes in Dao and Service hierarchy (add new abstract methods); 4. change primitive types in entity fields to their wraappers.

+++ FinalProject_LikeIT/src/main/java/com/kolyadko/likeit/command/impl/actioncommand/CreateCommentCommand.java
@@ -40,7 +40,7 @@ public class CreateCommentCommand extends ActionCommand {
                     CommentService commentService = new CommentService();
                     Comment comment = new Comment();
                     comment.setAuthorId(RequestContentUtil.getCurrentUserLogin(content));
-                    comment.setQuestionId(Integer.parseInt(container.getQuestionId()));
+                    comment.setQuestionId(Long.parseLong(container.getQuestionId()));
                     comment.setText(container.getText());
 
                     if (commentService.create(comment)) {
